requirements.txt
streamlit>=1.29.0

torch>=2.0.0
torchvision>=0.15.0

numpy>=1.23.0
pillow>=9.5.0
opencv-python>=4.8.0

scipy>=1.10.0
pandas>=2.0.0

reportlab>=4.0.0

matplotlib>=3.7.0

tqdm>=4.65.0
pymongo>=4.6.0
bcrypt>=4.1.0
python-dotenv>=1.0.0

reportlab>=4.0.0
